feat(154) Move Trip Search stuff to a dedicated concern
